requirements.txt
albumentations==1.1.0
neptune_client==0.16.7
numpy==1.21.2
opencv_python==4.5.5.64
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
